Necessary changes due to moving net package
