+++ dht_net.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/libp2p/go-libp2p-core/network"
 
-	"github.com/libp2p/go-libp2p-kad-dht/internal/net"
 	"github.com/libp2p/go-libp2p-kad-dht/metrics"
+	"github.com/libp2p/go-libp2p-kad-dht/net"
 	pb "github.com/libp2p/go-libp2p-kad-dht/pb"
 
 	"github.com/libp2p/go-msgio"
